Added Swiftlint (Linted)

+++ insights-for-instagram/App/AppDataStore.swift
@@ -12,9 +12,9 @@ import UIKit
 class AppDataStore {
     
     // Returns the top n most liked
-    class func getMostLiked (with limit:Int ) -> [InstagramMedia]{
+    class func getMostLiked (with limit: Int ) -> [InstagramMedia] {
         // paginating behavior isn’t necessary at all: https://realm.io/docs/swift/latest/#limiting-results
-        let realm = try! Realm()
+        guard let realm = try? Realm() else { return [] }
         let medias = realm.objects(InstagramMedia.self).sorted(byKeyPath: "commentsCount", ascending: false)
         if medias.count > limit {
             return Array(medias[0...limit])
@@ -23,18 +23,17 @@ class AppDataStore {
     }
     
     // Returns last (n) weeks posted media
-    class func getLastWeeksPosted (weeks:Int) -> [InstagramMedia]{
-        let realm = try! Realm()
-        let fromDate = Calendar.current.date(byAdding: .day, value: -(7 * weeks), to: Date())! as NSDate
-        let predicate = NSPredicate(format: "createdTime > %@", fromDate)
+    class func getLastWeeksPosted (weeks: Int) -> [InstagramMedia] {
+        guard let realm = try? Realm() else { return [] }
+        guard let fromDate = Calendar.current.date(byAdding: .day, value: -(7 * weeks), to: Date()) else { return [] }
+        let predicate = NSPredicate(format: "createdTime > %@", fromDate as CVarArg)
         let medias = realm.objects(InstagramMedia.self).sorted(byKeyPath: "createdTime", ascending: false).filter(predicate)
         return Array(medias)
     }
     
     // Returns best Engagement
-    class func getBestEngagement (with limit:Int) -> [InstagramMedia] {
-        // paginating behavior isn’t necessary at all: https://realm.io/docs/swift/latest/#limiting-results
-        let realm = try! Realm()
+    class func getBestEngagement (with limit: Int) -> [InstagramMedia] {
+        guard let realm = try? Realm() else { return [] }
         let medias = realm.objects(InstagramMedia.self).sorted(byKeyPath: "engagementCount", ascending: false)
         if medias.count > limit {
             return Array(medias[0...limit])
@@ -44,27 +43,29 @@ class AppDataStore {
     
     // Returns the oldest media stored locally
     class func getInstagramMediaIndex() -> (offset: String?, count: Int) {
-        let realm = try! Realm()
+        guard let realm = try? Realm() else { return (nil, 0) }
         let entries = realm.objects(InstagramMedia.self).sorted(byKeyPath: "createdTime", ascending: true)
-        if entries.count > 0 {
-            let offset = entries[0]["id"] as! String
+        if entries.isEmpty == false {
+            guard let offset = entries[0]["id"] as? String else { return (nil, 0) }
             let count = entries.count
             return (offset, count)
-        }else {
+        } else {
             return (nil, 0)
         }
     }
     
     // Creates/updates media
-    class func importInstagramMedia(instagramMedia: [[String: Any]], completion: @escaping () -> Void){
+    class func importInstagramMedia(instagramMedia: [[String: Any]], completion: @escaping () -> Void) {
         DispatchQueue.global().async {
-            let realm = try! Realm()
+            guard let realm = try? Realm() else { return }
             realm.beginWrite()
             for media in instagramMedia {
-                let instagramMedia = InstagramMedia(JSON: media)
-                realm.add(instagramMedia!, update: true)
+                guard let instagramMedia = InstagramMedia(JSON: media) else {
+                    continue
+                }
+                realm.add(instagramMedia, update: true)
             }
-            try! realm.commitWrite()
+            try? realm.commitWrite()
             DispatchQueue.main.async {
                 completion()
             }
@@ -73,8 +74,8 @@ class AppDataStore {
     
     class func deleteAll() {
         AppUserAccount().name = nil
-        let realm = try! Realm()
-        try! realm.write {
+        guard let realm = try? Realm() else { return }
+        try? realm.write {
             realm.deleteAll()
         }
         NotificationCenter.default.post(name: AppConfiguration.DefaultsNotifications.reload, object: nil)
